refactor: 87 - rename file with same address emitter logic
