+++ src/main.ts
@@ -1,10 +1,13 @@
 import './style.css';
-import { appState } from './app/app-state';
-import { renderPage } from './app/router/render-page';
-import { router } from './app/router/router';
+// import { appState } from './app/app-state';
+// import { renderPage } from './app/router/render-page';
+// import { router } from './app/router/router';
+import { RegistrationPage } from './pages/registration';
 
-(function (): void {
-  router();
-  renderPage(appState.currentPage);
-  globalThis.location.hash = `#${appState.currentPage}`;
-})();
+// (function (): void {
+//   router();
+//   renderPage(appState.currentPage);
+//   globalThis.location.hash = `#${appState.currentPage}`;
+// })();
+
+RegistrationPage();
